+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-git # Madara Changelog
+git # Deoxys Changelog
 
 ## Next release
 
+- fix(workflows): Fix deoxys CI
 - feat(rpc): add_invoke_tx, add_deploy_account_tx, add_declare_tx
 - feat(rpc): tx_receipt, re-execute tx
 - feat(script): added CI scripts for starting Deoxys and comparing JSON RPC